workers/pyproject.toml
[project]
name = "axiom-workers"
version = "0.1.0"
description = "Axiom Design Engine - Celery Workers"
requires-python = ">=3.11"
dependencies = [
    # Task Queue
    "celery[redis]>=5.3.6",
    "redis>=5.0.1",
    "kombu>=5.3.4",
    
    # AI/ML Dependencies
    "torch>=2.1.2",
    "torchvision>=0.16.2",
    "transformers>=4.36.2",
    "diffusers>=0.25.0",
    "accelerate>=0.25.0",
    "safetensors>=0.4.1",
    
    # Image/Video Processing
    "pillow>=10.2.0",
    "opencv-python-headless>=4.9.0",
    "numpy>=1.26.3",
    
    # HTTP Client
    "httpx>=0.26.0",
    "websockets>=12.0",
    "aiofiles>=23.2.1",
    
    # Storage
    "boto3>=1.34.0",
    
    # Utilities
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "orjson>=3.9.10",
    
    # Observability
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-celery>=0.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.11",
    "mypy>=1.8.0",
]

gpu = [
    # CUDA-specific packages
    "xformers>=0.0.23",
    "triton>=2.1.0",
]

model3d = [
    # 3D model generation
    "trimesh>=4.0.0",
    "rembg>=2.0.50",
]

[project.scripts]
axiom-worker = "workers.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
